% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{chisq_distance}
\alias{chisq_distance}
\title{Matrix of chi-square distances}
\usage{
chisq_distance(table, indices = seq_len(dim(table)[dimension]),
  dimension = 1)
}
\arguments{
\item{table}{object of class "table".}

\item{indices}{integer vector. Which rows or columns should the function calculate distances between.\
Default is all the row or all the columns depending on what is chosen in the argument "dimension".}

\item{dimension}{integer. Whether to calculate distances between rows (1) or columns (2). Default is 1.}
}
\value{
Matrix of size length(indices) x length(indices) containing chi-square distances between selected
rows or columns.
}
\description{
Calculates chi-square distances between selected rows or columns of the contingency table.
}
\examples{
data(ksarakil)
chisq_distance(ksarakil)
}
