% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{cluster_distance}
\alias{cluster_distance}
\title{Matrix of Ward's distances between clusters}
\usage{
cluster_distance(table, clusters, dimension = 1)
}
\arguments{
\item{table}{object of class "table".}

\item{clusters}{list of integer vectors. Each vector should define a cluster
by specifing row or column indices of its memebrs. Clusters must not overalap.}

\item{dimension}{integer. Whether to use rows (1) or columns (2). Default is 1.}
}
\value{
Matrix of size length(clusters) x length(clusters) containing distances between selected clusters of
rows or columns.
}
\description{
Calculates chi-square distances between selected rows or columns of the contingency table.
}
\examples{
data(israeli_survey)
cluster_distance(israeli_survey, as.list(seq_len(nrow(israeli_survey))), 1)

cluster_distance(israeli_survey, list(1, 2, c(3, 5), c(4, 6, 7), 8), 1)
}
