% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clusters2list}
\alias{clusters2list}
\alias{kmeans2list}
\title{Refactor cluster memebership}
\usage{
clusters2list(vec)

kmeans2list(kmeans_result)
}
\arguments{
\item{vec}{vector of cluster membership indices.}

\item{kmeans_result}{object of class kmeans}
}
\value{
list of vectors of cluster members
}
\description{
Turn clsuter membershiip vector either from kmeans or cutree into list of vectors of cluster members
}
\examples{
clusters2list(c(1, 1, 2, 3, 1, 2, 3))

clusters <- kmeans_table(israeli_survey, dimesnion = 1, k = 4)
kmeans2list(clusters)
}
