% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{distance_bound}
\alias{distance_bound}
\title{Upper bound for distance between clusters}
\usage{
distance_bound(table, alpha = 0.05, options = list(samples = 1000, seed
  = 42))
}
\arguments{
\item{table}{object of class table. The contingency table to calculate the upper bound for.}

\item{alpha}{double. The returned value is 1-alpha quantile of the underlying distribution. Default is 0.05.}

\item{options}{list. Two elements taken into accout is "samples", which says how many repetitions to use
in simulating distribution (default: 1000) and "seed" used in sampling if not null (default: NULL).}
}
\value{
The 1-alpha quantile of the distribution of maximal estimated squared contrast
in tables of the size like the given table under the null hypothesis of independence or homogeneity.
}
\description{
Under the null hypothesis of independence or homogeneity,
the maximal squared contrast in rows or columns of contingency table,
which can be interpreted as distance between rows is asympotically distributed as
the largest eigenvalue of the Wishart matrix W(I_min(I,J), max(I,J)), where I and J
are dimensions of the table.
}
\examples{
distance_bound(israeli_survey, options = list(samples = 1000, seed = 42))
}
