% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{gap_statistic}
\alias{gap_statistic}
\title{The gap statistic Work in progress}
\usage{
gap_statistic(table, dimension, clusters, func, scheme, B = 500)
}
\arguments{
\item{table}{object of class "table".}

\item{dimension}{integer. Whether to calculate distances between rows (1) or columns (2). Default is 1.}

\item{clusters}{list of integer vectors. Each vector should define a cluster
by specifing row or column indices of its memebrs. Clusters must not overalap.}

\item{func}{character. Name of clustering function. Currently "kmeans" and "hclust" are supported.}

\item{scheme}{character. Name of the null sampling scheme tables are genereated from. Can be either..}

\item{B}{integer. Number of sampled tables.}
}
\value{
list with elements: gap - value of the statistic and s -sclaled standard deviation
}
\description{
Quantifies clustering quailty comparing it to siumulated clusters.
See Section 4.1 of the thesis for further details.
}
\examples{
clusters <- kmeans_table(israeli_survey, dimesnion = 1, k = 4)
gap_statistic(israeli_survey, dimension = 1, kmeans2list(clusters), "kmeans", "multinomial", 100)
}
