% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{generalized_distance}
\alias{generalized_distance}
\title{Generalized distance between clusters}
\usage{
generalized_distance(table, clusters, dimension = 1)
}
\arguments{
\item{table}{object of class "table".}

\item{clusters}{list of integer vectors. Each vector should define a cluster
by specifing row or column indices of its memebrs. Clusters must not overalap.}

\item{dimension}{integer. Whether to calculate distances between rows (1) or columns (2). Default is 1.}
}
\value{
double. Value of generalized distance between clusters
}
\description{
Generalized distance between clusters
}
\examples{
data(israeli_survey)
generalized_distance(israeli_survey, as.list(seq_len(nrow(israeli_survey))), 1)

generalized_distance(israeli_survey, list(1, 2, c(3, 5), c(4, 6, 7)), 1)
}
