% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{hclust_table}
\alias{hclust_table}
\title{Hierarchical clustering of contingency tables}
\usage{
hclust_table(table, dimension = 1)
}
\arguments{
\item{table}{object of class "table".}

\item{dimension}{integer. Whether to calculate distances between rows (1) or columns (2). Default is 1.}
}
\value{
object of class hclust.
}
\description{
Applies Ward's agglomerative clustering to the rows or columns of the contingecy table.
}
\examples{
hclust_table(ksarakil)
hclust_table(israeli_survey, dimension = 2)
}
