% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{israeli_survey}
\alias{israeli_survey}
\title{Principal worries of 1554 Israeli survey participants}
\format{The 8x5 contingency table}
\source{
Gutttman
}
\usage{
israeli_survey
}
\description{
Principal worries of 1554 Israeli survey participants
}
\keyword{datasets}
