% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kmeans_table}
\alias{kmeans_table}
\title{K-means on correspondence analysis}
\usage{
kmeans_table(table, dimension = 1, k, ...)
}
\arguments{
\item{table}{object of class "table".}

\item{dimension}{integer. Whether to use rows (1) or columns (2). Default is 1.}

\item{k}{integer. number of clusters}

\item{...}{arguments passed to kmeans.}
}
\value{
object of class k-means with additional information about clustewr membership of items
classified in the contingecy table and about sum of squares in low-dimensional space used in clustering.
}
\description{
Performs weighted k-means clustering on k-1 in fir k-1 principal coordinates (or all of them),
obtained from correspondence analysis of contingency table.
}
\examples{
kmeans_table(ksarakil, 1, 5)
kmeans_table(israeli_survey, 2, 2)
}
