% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ksarakil}
\alias{ksarakil}
\title{Lithic tools from Ksar-Akil excavations}
\format{The 10x5 contingency table.}
\source{
Alesi
}
\usage{
ksarakil
}
\description{
Lithic tools from Ksar-Akil excavations
}
\keyword{datasets}
