% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{silhouette}
\alias{silhouette}
\title{Silhouette score}
\usage{
silhouette(table, dimension, clusters)
}
\arguments{
\item{table}{object of class "table".}

\item{dimension}{integer. Whether to use rows (1) or columns (2). Default is 1.}

\item{clusters}{list of integer vectors. Each vector should define a cluster
by specifing row or column indices of its memebrs. Clusters must not overalap.}
}
\value{
list with two elements. First element is a list with silhouette scores for each profile didvided
into clutsers, second one is a weighted mean of those scores.
}
\description{
Calculates silhouette scores for clustered profiles and mean score for the whole clustering result.
}
\examples{
clusters <- clusters2list(stats::cutree(hclust_table(ksarakil, 1), k = 4))
silhouette(ksarakil, 1, clusters)
}
